// frontend/src/gen/branch_connect.ts
// @generated by protoc-gen-connect-es v1.7.0 with parameter "target=ts,import_extension=none"
// @generated from file branch.proto (package branch, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { Branch, CreateBranchRequest, DeleteBranchRequest, GetBranchRequest, ListBranchesRequest, ListBranchesResponse, UpdateBranchRequest } from "./branch_pb";
import { MethodKind } from "@bufbuild/protobuf";
import { SuccessResponse } from "./common_pb";

/**
 * @generated from service branch.BranchService
 */
export const BranchService = {
  typeName: "branch.BranchService",
  methods: {
    /**
     * @generated from rpc branch.BranchService.ListBranches
     */
    listBranches: {
      name: "ListBranches",
      I: ListBranchesRequest,
      O: ListBranchesResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc branch.BranchService.CreateBranch
     */
    createBranch: {
      name: "CreateBranch",
      I: CreateBranchRequest,
      O: Branch,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc branch.BranchService.UpdateBranch
     */
    updateBranch: {
      name: "UpdateBranch",
      I: UpdateBranchRequest,
      O: Branch,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc branch.BranchService.DeleteBranch
     */
    deleteBranch: {
      name: "DeleteBranch",
      I: DeleteBranchRequest,
      O: SuccessResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc branch.BranchService.GetBranch
     */
    getBranch: {
      name: "GetBranch",
      I: GetBranchRequest,
      O: Branch,
      kind: MethodKind.Unary,
    },
  }
} as const;